refactor: InviteAdapter 패키지 이동
